Don't take a detour via HighlightingAssets to build assets
Move code to build assets to its own file. That results in better modularity.

It also allows us to simplify HighlitingAssets a lot, since it will now always
be initialized with a SerializedSyntaxSet (see next commit).
